refactor(feeds): db methods for find

+++ internal/events/subscribers.go
@@ -11,6 +11,7 @@ import (
 	"github.com/autobrr/autobrr/internal/logger"
 	"github.com/autobrr/autobrr/internal/notification"
 	"github.com/autobrr/autobrr/internal/release"
+	"github.com/autobrr/autobrr/pkg/errors"
 
 	"github.com/asaskevich/EventBus"
 	"github.com/rs/zerolog"
@@ -40,13 +41,13 @@ func NewSubscribers(log logger.Logger, eventbus EventBus.Bus, feedSvc feed.Servi
 }
 
 func (s Subscriber) Register() {
-	s.eventbus.Subscribe(domain.EventReleaseStoreActionStatus, s.releaseActionStatus)
-	s.eventbus.Subscribe(domain.EventReleasePushStatus, s.releasePushStatus)
-	s.eventbus.Subscribe(domain.EventNotificationSend, s.sendNotification)
-	s.eventbus.Subscribe(domain.EventIndexerDelete, s.deleteIndexer)
+	s.eventbus.Subscribe(domain.EventReleaseStoreActionStatus, s.handleReleaseActionStatus)
+	s.eventbus.Subscribe(domain.EventReleasePushStatus, s.handleReleasePushStatus)
+	s.eventbus.Subscribe(domain.EventNotificationSend, s.handleSendNotification)
+	s.eventbus.Subscribe(domain.EventIndexerDelete, s.handleIndexerDelete)
 }
 
-func (s Subscriber) releaseActionStatus(actionStatus *domain.ReleaseActionStatus) {
+func (s Subscriber) handleReleaseActionStatus(actionStatus *domain.ReleaseActionStatus) {
 	s.log.Trace().Str("event", domain.EventReleaseStoreActionStatus).Msgf("store action status: '%+v'", actionStatus)
 
 	err := s.releaseSvc.StoreReleaseActionStatus(context.Background(), actionStatus)
@@ -55,7 +56,7 @@ func (s Subscriber) releaseActionStatus(actionStatus *domain.ReleaseActionStatus
 	}
 }
 
-func (s Subscriber) releasePushStatus(actionStatus *domain.ReleaseActionStatus) {
+func (s Subscriber) handleReleasePushStatus(actionStatus *domain.ReleaseActionStatus) {
 	s.log.Trace().Str("event", domain.EventReleasePushStatus).Msgf("events: 'release:push' '%+v'", actionStatus)
 
 	if err := s.releaseSvc.StoreReleaseActionStatus(context.Background(), actionStatus); err != nil {
@@ -63,25 +64,33 @@ func (s Subscriber) releasePushStatus(actionStatus *domain.ReleaseActionStatus) 
 	}
 }
 
-func (s Subscriber) sendNotification(event *domain.NotificationEvent, payload *domain.NotificationPayload) {
+func (s Subscriber) handleSendNotification(event *domain.NotificationEvent, payload *domain.NotificationPayload) {
 	s.log.Trace().Str("event", domain.EventNotificationSend).Msgf("send notification events: '%v' '%+v'", *event, payload)
 
 	s.notificationSvc.Send(*event, *payload)
 }
 
-// deleteIndexer handle feed cleanup via event because feed service can't be imported in indexer service
-func (s Subscriber) deleteIndexer(indexerID int) {
-	s.log.Trace().Str("event", domain.EventIndexerDelete).Msgf("events: 'indexer:delete' '%d'", indexerID)
+// handleIndexerDelete handle feed cleanup via event because feed service can't be imported in indexer service
+func (s Subscriber) handleIndexerDelete(indexer *domain.Indexer) {
+	s.log.Trace().Str("event", domain.EventIndexerDelete).Msgf("events: 'indexer:delete' '%d'", indexer.ID)
 
 	ctx := context.Background()
 
-	feedItem, err := s.feedSvc.FindOne(ctx, domain.FindOneParams{IndexerID: indexerID})
-	if err != nil {
-		s.log.Error().Err(err).Msgf("events: 'indexer:delete' error, could not find feed with indexer id: %d", indexerID)
-		return
-	}
+	if indexer.ImplementationIsFeed() {
+		feedItem, err := s.feedSvc.FindOne(ctx, domain.FindOneParams{IndexerID: int(indexer.ID)})
+		if err != nil {
+			if errors.Is(err, domain.ErrRecordNotFound) {
+				return
+			}
 
-	if err := s.feedSvc.Delete(ctx, feedItem.ID); err != nil {
-		s.log.Error().Err(err).Msgf("events: 'indexer:delete' error, could not delete feed with id: %d", feedItem.ID)
+			s.log.Error().Err(err).Msgf("events: 'indexer:delete' error, could not find feed with indexer id: %d", indexer.ID)
+			return
+		}
+
+		if err := s.feedSvc.Delete(ctx, feedItem.ID); err != nil {
+			s.log.Error().Err(err).Msgf("events: 'indexer:delete' error, could not delete feed with id: %d", feedItem.ID)
+		}
+
+		s.log.Debug().Msgf("successfully removed feed: %s", feedItem.Name)
 	}
 }
